Looks the opposite of starting state is never allowed which to me is not explained properly.
The exercise states "To help, the tests provide you with which bucket to fill first. That means, when starting with the larger bucket full, you are NOT allowed at any point to have the smaller bucket full and the larger bucket empty (aka, the opposite starting point); that would defeat the purpose of comparing both approaches!", which indicates that the opposite state is not allowed only when the start bucket is the larger bucket.